stop wrapping types with sqlx::types::Json
We now implement the sqlx traits directly on our structs, so
no need to wrap them in sqlx::types::Json anymore.

Had to introduce a new types, JVec<T>, because we cannot implement
traits from other crates on a Vec<T> even if T is a local type :(

This now makes it possible to use derive(poemopen_api::Object) on
all our structs as well. Except on JVec<T> because the derive does
not support generics, so for JVec it is implemented manually.